TestCaseQasm implementation.
